BUGFIX:
By enabling the synch = ture in the ajax request, the tagfield didn't work well in case that the search in server side takes long time, it has 2 bad side effects, 1. a early request might return the response late, and hence, cover a later request's response. 2. the client side (browser side) fill with spining request. In addition, currently tagfield doesn't work with 'Backspace' key, this is because the delay is set to 300 in the jquery.tags.js for some reason. So we decide to apply a patch tag_asych_request to this field.

This is submission for preparing to apply the patch, mainly a partial rollback of r67871