Fixing setValue null value bug
On occasions when a form is submitted `StringTagField` throws an error saying `ERROR [Warning]: array_filter() expects parameter 1 to be array, null given` at the end of the `setValue` function. 

I have added some code that checks if `$value` is `null`. If it is `null` it sets `$value` to an empty array to prevent the error from occurring.